+++ lib/FluxDispatcher.js
@@ -1,4 +1,8 @@
-enyo.singleton({
+var
+	kind = require('./kind'),
+	CoreObject = require('./CoreObject');
+
+module.exports = kind.singleton({
 	/**
 	*
     *
@@ -8,7 +12,7 @@ enyo.singleton({
 	*/
 	name: 'enyo.FluxDispatcher',
 
-	kind: enyo.Object,
+	kind: Object,
 
 	/**
 	* These subscriptions to a stores state updates
